requirements.txt
python-telegram-bot
edge-tts
yt-dlp
flask
